Added most auth to endpoints

+++ app-api/Controllers/AutomationController.cs
@@ -1,5 +1,7 @@
+using System.Security.Claims;
 using app_api.Models;
 using AutoMapper;
+using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.EntityFrameworkCore;
 
@@ -9,39 +11,59 @@ public class AutomationController : Controller
 {
     private readonly IAutomationRepository automationRepository;
     private readonly IMapper mapper;
+    private readonly IWorkspaceRepository workspaceRepository;
 
-    public AutomationController(IAutomationRepository automationRepository, IMapper mapper)
+    public AutomationController(IAutomationRepository automationRepository, IMapper mapper, IWorkspaceRepository workspaceRepository)
     {
         this.automationRepository = automationRepository;
         this.mapper = mapper;
+        this.workspaceRepository = workspaceRepository;
     }
 
     [HttpGet]
     [Route("{workspaceId:guid}")]
+    [Authorize]
     public async Task<IActionResult> GetAutomationsByWorkspaceId(Guid workspaceId)
     {
-        var workspaceAutomations = await automationRepository.GetAutomationsByWorkspaceIdAsync(workspaceId);
+        var idInToken = Guid.Parse(User.FindFirstValue(ClaimTypes.NameIdentifier)!);
 
-        if (workspaceAutomations == null)
+        var workspace = await workspaceRepository.GetWorkspaceByIdAsync(workspaceId);
+
+        if (workspace == null)
         {
             return NotFound("Workspace not found!");
         }
+        else if(workspace.OwnerId != idInToken)
+        {
+            return Forbid();
+        }
+        
+        var workspaceAutomations = await automationRepository.GetAutomationsByWorkspaceIdAsync(workspaceId);
 
         return Ok(mapper.Map<List<AutomationDTO>>(workspaceAutomations));
     }
 
     [HttpPost]
     [ValidateModel]
+    [Authorize]
     public async Task<IActionResult> CreateAutomation([FromBody] CreateAutomationDTO createAutomationDTO)
     {
-        var newAutomation = mapper.Map<Automation>(createAutomationDTO);
+        var idInToken = Guid.Parse(User.FindFirstValue(ClaimTypes.NameIdentifier)!);
 
-        newAutomation = await automationRepository.CreateAutomationAsync(newAutomation);
+        var workspace = await workspaceRepository.GetWorkspaceByIdAsync(createAutomationDTO.WorkspaceId);
 
-        if (newAutomation == null)
+        if (workspace == null)
         {
             return NotFound("Workspace not found!");
         }
+        else if(workspace.OwnerId != idInToken)
+        {
+            return Forbid();
+        }
+
+        var newAutomation = mapper.Map<Automation>(createAutomationDTO);
+
+        newAutomation = await automationRepository.CreateAutomationAsync(newAutomation);
 
         var returnAuto = mapper.Map<AutomationDTO>(newAutomation);
 
@@ -51,29 +73,47 @@ public class AutomationController : Controller
     [HttpPut]
     [ValidateModel]
     [Route("{automationId:guid}")]
+    [Authorize]
     public async Task<IActionResult> UpdateAutomation([FromRoute] Guid automationId, [FromBody] UpdateAutomationDTO updateAutomationDTO)
     {
+        var idInToken = Guid.Parse(User.FindFirstValue(ClaimTypes.NameIdentifier)!);
+
+        var automation = await automationRepository.GetAutomationByIdAsync(automationId);
+
+        if (automation == null)
+        {
+            return NotFound("Automation not found!");
+        }
+        else if(automation.Workspace.OwnerId != idInToken)
+        {
+            return Forbid();
+        }
+
         var automationToUpdate = mapper.Map<Automation>(updateAutomationDTO);
 
         automationToUpdate = await automationRepository.UpdateAutomationAsync(automationId, automationToUpdate);
-
-        if (automationToUpdate == null)
-        {
-            return NotFound("Automation doesn't exist!");
-        }
 
         return Ok(mapper.Map<AutomationDTO>(automationToUpdate));
     }
     [HttpDelete]
     [Route("{automationId:guid}")]
+    [Authorize]
     public async Task<IActionResult> DeleteAutomation([FromRoute] Guid automationId)
     {
-        var automationToDelete = await automationRepository.DeleteAutomationAsync(automationId);
+        var idInToken = Guid.Parse(User.FindFirstValue(ClaimTypes.NameIdentifier)!);
 
-        if (automationToDelete == null)
+        var automation = await automationRepository.GetAutomationByIdAsync(automationId);
+
+        if (automation == null)
         {
-            return NotFound("Automation doesn't exist!");
+            return NotFound("Automation not found!");
         }
+        else if(automation.Workspace.OwnerId != idInToken)
+        {
+            return Forbid();
+        }
+
+        await automationRepository.DeleteAutomationAsync(automationId);
 
         return Ok();
     }
